first changes
